csi_loc/utils/config.py
import os
import yaml
from typing import Any, Dict


def load_config(path: str) -> Dict[str, Any]:
    with open(path, "r", encoding="utf-8") as f:
        cfg = yaml.safe_load(f)
    return cfg


def ensure_dir(path: str) -> None:
    os.makedirs(path, exist_ok=True)
