requirements.txt
torch>=2.2.0
numpy>=1.24
PyYAML>=6.0
scikit-learn>=1.2
tqdm>=4.66
